+++ base_accounts/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (2, 3, 1)
+VERSION = (2, 3, 2)
 __version__ = '.'.join(map(str, VERSION))