+++ base_accounts/forms.py
@@ -2,22 +2,45 @@ from django import forms
 from django.contrib.auth import authenticate, login, get_user_model
 from django.utils.translation import ugettext_lazy as _
 
+from base_accounts.utils import create_email_user
+
 
 class SignupForm(forms.Form):
     full_name = forms.CharField(label=_('full name'))
     email = forms.EmailField(label=_('email'))
     password = forms.CharField(label=_('password'), widget=forms.PasswordInput)
     tos = forms.BooleanField(label=_('I accept the terms of service'))
+    user_model = get_user_model()
+
+    def __init__(self, *args, **kwargs):
+        request = kwargs.pop('request', None)
+        self.request = request
+        return super(SignupForm, self).__init__(*args, **kwargs)
 
     def clean_email(self, *args, **kwargs):
         data = self.cleaned_data['email']
-        user_model = get_user_model()
         try:
-            user_model.objects.get(email=data)
-        except user_model.DoesNotExist:
+            self.user_model.objects.get(email=data)
+        except self.user_model.DoesNotExist:
             return data
         else:
             raise forms.ValidationError(_("Email is already being used by another user"))
+
+    def save(self, *args, **kwargs):
+
+        # Get model fields
+        full_name = self.cleaned_data.get('full_name').lower()
+        email = self.cleaned_data.get('email').lower()
+        password = self.cleaned_data.get('password')
+
+        # Create new user
+        user = create_email_user(email, password, self.user_model, **{'name': full_name})
+
+        # Authenticate and login user
+        user = authenticate(username=user.username, password=password)
+        login(self.request, user)
+
+        return user
 
 
 class LoginForm(forms.Form):
@@ -48,6 +71,7 @@ class LoginForm(forms.Form):
 
 class UpdateEmailForm(forms.Form):
     email = forms.EmailField(required=True)
+    user_model = get_user_model()
 
     def __init__(self, *args, **kwargs):
         request = kwargs.pop('request', None)
@@ -56,10 +80,9 @@ class UpdateEmailForm(forms.Form):
 
     def clean_email(self, *args, **kwargs):
         data = self.cleaned_data['email']
-        user_model = get_user_model()
         try:
-            user_model.objects.exclude(id=self.user.id).get(email=data)
-        except user_model.DoesNotExist:
+            self.user_model.objects.exclude(id=self.user.id).get(email=data)
+        except self.user_model.DoesNotExist:
             return data
         else:
             raise forms.ValidationError(_("Email is already being used by another user"))
@@ -68,6 +91,7 @@ class UpdateEmailForm(forms.Form):
 class UpdatePasswordForm(forms.Form):
     password1 = forms.CharField(label=_('new password'), widget=forms.PasswordInput,)
     password2 = forms.CharField(label=_('new password (confirm)'), widget=forms.PasswordInput,)
+    user_model = get_user_model()
 
     def __init__(self, *args, **kwargs):
         request = kwargs.pop('request', None)
@@ -83,5 +107,4 @@ class UpdatePasswordForm(forms.Form):
 
     def save(self):
         self.user.set_password(self.cleaned_data['password1'])
-        user_model = get_user_model()
-        user_model.objects.filter(pk=self.user.pk).update(password=self.user.password)
+        self.user_model.objects.filter(pk=self.user.pk).update(password=self.user.password)